fix: create history.interface file
