+++ src/interfaces/history.interface.ts
@@ -1,0 +1,12 @@
+export interface IntAttributtes {
+  place: string;
+  city: string;
+  country: string;
+  coordinates: [number, number];
+}
+
+export interface IntSearch {
+  type: string;
+  id: number;
+  attributes: IntAttributtes;
+}
